separate composables and add previews
