+++ app/src/main/java/com/example/compasstest/ui/Footer.kt
@@ -1,0 +1,26 @@
+package com.example.compasstest.ui
+
+import androidx.compose.foundation.layout.Column
+import androidx.compose.foundation.layout.fillMaxWidth
+import androidx.compose.foundation.layout.padding
+import androidx.compose.material3.Button
+import androidx.compose.material3.Divider
+import androidx.compose.material3.Text
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.unit.dp
+
+@Composable
+fun Footer(onClick: () -> Unit) {
+    Column {
+        Divider()
+        Button(
+            modifier = Modifier
+                .fillMaxWidth()
+                .padding(16.dp),
+            onClick = onClick,
+        ) {
+            Text("Show every tenth character")
+        }
+    }
+}
