+++ app/src/main/java/com/example/compasstest/ui/MainActivity.kt
@@ -4,22 +4,8 @@ import android.os.Bundle
 import androidx.activity.ComponentActivity
 import androidx.activity.compose.setContent
 import androidx.activity.viewModels
-import androidx.compose.foundation.layout.Column
-import androidx.compose.foundation.layout.fillMaxSize
-import androidx.compose.foundation.layout.fillMaxWidth
-import androidx.compose.foundation.layout.padding
-import androidx.compose.foundation.rememberScrollState
-import androidx.compose.foundation.verticalScroll
-import androidx.compose.material3.Button
-import androidx.compose.material3.Divider
-import androidx.compose.material3.Text
-import androidx.compose.runtime.Composable
 import androidx.compose.runtime.collectAsState
 import androidx.compose.runtime.getValue
-import androidx.compose.ui.Modifier
-import androidx.compose.ui.unit.dp
-import androidx.constraintlayout.compose.ConstraintLayout
-import com.example.compasstest.ui.theme.CompassTestTheme
 import dagger.hilt.android.AndroidEntryPoint
 
 @AndroidEntryPoint
@@ -29,42 +15,12 @@ class MainActivity : ComponentActivity() {
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
 
-        setContent { MainContent() }
-    }
-
-    @Composable
-    fun MainContent() {
-        val everyTenthCharacterResponse by viewModel.response.collectAsState()
-
-        CompassTestTheme {
-            Column(modifier = Modifier.fillMaxSize()) {
-                val scrollState = rememberScrollState()
-
-                Text(
-                    text = everyTenthCharacterResponse,
-                    modifier = Modifier
-                        .weight(1f)
-                        .verticalScroll(scrollState)
-                        .padding(16.dp)
-                )
-
-                Footer()
-            }
-        }
-    }
-
-    @Composable
-    fun Footer() {
-        Column {
-            Divider()
-            Button(
-                modifier = Modifier
-                    .fillMaxWidth()
-                    .padding(16.dp),
-                onClick = { viewModel.getEveryTenthCharacter() },
-            ) {
-                Text("Show every tenth character")
-            }
+        setContent {
+            val everyTenthCharacterResponse by viewModel.response.collectAsState()
+            MainContent(
+                firstText = everyTenthCharacterResponse,
+                onFooterClick = { viewModel.getEveryTenthCharacter() }
+            )
         }
     }
 }
